requirements.txt
pandas
numpy
streamlit
plotly
pillow

